backend/requirements.txt
flask
flask-cors
python-dotenv
langchain
langchain_community
langchain_ollama
